refactor: extract shared template helpers into keeperhub/lib/template-helpers
Move 11 duplicated functions and 3 types from template-code-editor,
sql-template-editor, and template-autocomplete into a shared module.
Eliminates ~400 lines of duplicated field extraction and node display
logic across the three editor components.
